+++ decedric/__init__.py
@@ -1,0 +1,9 @@
+from decedric.core_simple import Variable
+from decedric.core_simple import Function
+from decedric.core_simple import using_config
+from decedric.core_simple import no_grad
+from decedric.core_simple import as_array
+from decedric.core_simple import as_variable
+from decedric.core_simple import setup_variable
+
+setup_variable()
